making everthing ready for the First Lesson
