+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/opmode/teleop/TeleOPLesson.java
@@ -1,0 +1,13 @@
+//package org.firstinspires.ftc.teamcode.robot.opmode.teleop;
+//
+////TODO uncommeting:Everything in this class is commented (everthing is grey) so it does not work. To uncomment everything, Select everything and press 'CTRL' and '/' at the same time.
+//
+////TODO goal: The Drivetrain needs to drive at the speed that is given a the controller.
+//
+////TODO Exercise: Make sure that the Function of 'SubsystemLesson' works.
+//
+////ASKING FOR HELP IS NO SHAME!
+////TIP: Try to understand what is happening in 'TeleOpEexample'. If you don't know, why don't you ask?
+//
+//public class TeleOPLesson {
+//}
