--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/opmode/teleop/TeleOPLesson1_2.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.robot.opmode.teleop;
-
-public class TeleOPLesson1_2 {
-}
