+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystem/ExampleSubsystem.java
@@ -3,7 +3,6 @@ package org.firstinspires.ftc.teamcode.robot.subsystem;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-import com.qualcomm.robotcore.hardware.Servo;
 
 
 /**
