+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystem/SubsystemLesson.java
@@ -1,0 +1,15 @@
+//package org.firstinspires.ftc.teamcode.robot.subsystem;
+//
+////TODO uncommeting:Everything in this class is commented (everthing is grey) so it does not work. To uncomment everything, Select everything and press 'CTRL' and '/' at the same time.
+//
+////TODO goal: The Drivetrain needs to drive at the speed that is given by a controller.
+//
+////TODO Exercise: In this class, you need to make a function that turns 4 motors at the exact same speed at the exact same moment.
+//// These motors are the motors of the drivetrain.
+//// The Speed needs to be controller by a controller.
+//
+////ASKING FOR HELP IS NO SHAME!
+////TIP: Try to understand what is happening in 'ExampleSubsystem'. If you don't know, why don't you ask?
+//
+//public class SubsystemLesson {
+//}
