--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystem/SubsystemLesson1_2.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.robot.subsystem;
-
-public class SubsystemLesson1_2 {
-}
